Re-apply fixes to the "SIM network unlock" UI (which were lost in the big CDMA merge)
CL 138604 in cupcake fixed some problems with the "SIM network unlock"
UI (bug 1695133.)  That change was automerged to //branches/donutburger
in CL 139143, but it then got mostly undone when CL 144185 (the big CDMA
merge) went in.

This change manually re-applies the fixes from CL 138604,
this time in the donut git tree.

The fixes:
  - remove an unnecessary setMovementMethod(null) call
  - remove the "Dismiss" button
  - minor layout tweaks
  - minor comment improvements

TESTED: faked an EVENT_SIM_NETWORK_LOCKED event to bring up the
IccNetworkDepersonalizationPanel, verified the "dismiss" button is gone
and the soft keyboard comes up correctly.