Fix 1809590: text is truncated in a couple of "incoming call" menu items
The in-call screen has a couple of menu items with labels that are too
long for a single line.  So for those items we turn off the icon (which
leaves enough room for 2 lines) and use explicit '\n' chars in the
string resources to break the text into two lines.

Unfotunately this never actually worked :-(
The InCallMenuItemView class has always forced itself into single-line
mode, which causes the '\n' chars in the text to be ignored.

The fix: *don't* force single-line mode for menu items that don't have
an icon.  (If there's no icon, that means there *is* room for 2 lines of
text.)

TESTED: Confirmed that the multi-line menu items now look correct, and
the other in-call menu items are unchanged.