Fix NPE in CallCard.onQueryComplete (bug 1980378).
Apparently in some (very rare) cases it's possible for a Call to have no
connections by the time we try to update the CallCard.

I was never able to reproduce this, but I can at least fix
CallCard.onQueryComplete() to gracefully handle a null connection.

TESTED: lots of incoming and outgoing calls; confirmed no crashes, and
        that the info on the CallCard is still correct.