Check for phone state change.
There are two state change notifications - Call State change
and Phone State change. We register for Call State change,
and so when we get a notification Phone State might not have changed.
Hence, check for old Phone state change. This was causing a problem
with some headsets where we were sending updates twice.