This fixes an NULL pointer exception in Settings.
On sholes when opening the Settings -> Wireless controls -> Mobile newtorks
a NPE occurs because otherNetworkType is NULL. These two statements
aren't present in the master branch so I've removed as a first
guess on a solution. Will have Amith review this change to be
sure it isn't breaking something else.