Phone: hook up proximity sensor support (bug 1781059).
On devices that have a proximity sensor, the Phone app now holds a
PROXIMITY_SCREEN_OFF_WAKE_LOCK (see Change 8681) any time the phone is
offhook, regardless of what app is in the foreground.

TESTED:
 - On device #1 (with proximity sensor): screen turns off automatically
   when I hold it up to my face.  Verified this with both InCallScreen and
   Home app in foreground.
 - On device #2 (no proximity sensor): no new behavior.

Watch out: there are still some issues with how the PowerManager and the
lock screen interact while the PROXIMITY_SCREEN_OFF_WAKE_LOCK is held.
(In particular, if you go to some other app while a call is active, and
hold the phone up to your face, you'll be forced to unlock the device
*every* time you move the phone away from your face.)