Update silent-mode behavior and haptic feedback in the EmergencyDialer too.
Update the phone app's EmergencyDialer with a couple of changes that were
recently added to the regular dialer that's part of the Contacts app:

- Don't play touch tone sounds if the phone is in "silent mode".
  (This corresponds to changes https://android-git.corp.google.com/g/5438
  and https://android-git.corp.google.com/g/7963 for the regular dialer;
  see bug 1839556.)

- Add haptic feedback.  (This corresponds to change
  https://android-git.corp.google.com/g/7502 for the regular Dialer; see
  bug 1940121.)

The code changes here are *identical* to the corresponding TwelveKeyDialer
changes mentioned above.  (Yes, unfortunately there's a lot of code
duplicated between the TwelveKeyDialer and the EmergencyDialer.  But since
one's in apps/Contacts and the other is in apps/Phone, there's no easy way
to share common code.  I added a TODO comment mentioning this.)

Also, fix the EmergencyDialer's ToneGenerator to use
AudioManager.STREAM_DTMF, like the regular dialer does.