Force in-call UI to stay in portrait mode.
Currently, on devices with a sliding keyboard, the in-call screen (like
most other apps) switches automatically between landscape and portrait
based on the state of the keyboard.

But supporting this "feature" for the in-call UI turns out to require a
lot of design and engineering work(*) for very little benefit(**).

So let's make things simpler by using *only* portrait mode
for the in-call UI.

(*) The extra design / engineering work for this is pretty significant,
    since the layout of the in-call screen needs to be very different in the
    two modes, especially now that we have onscreen buttons on some devices.
    This causes extra work for testing as well.

(**) There's very little need for landscape mode in the first place, since
    hardly anybody talks on the phone with the keyboard open (simply because
    it's awkward to hold the phone that way.)  It *is* common to have the
    keyboard open while doing something else if you're using the
    speakerphone or bluetooth, but by that point you're not in the in-call
    UI any more.

    Basically the only reason we *do* switch to landscape mode now is that
    it looks goofy to have the screen in portrait mode while the keyboard is
    open.  But since it's awkward to talk on the phone with the keyboard
    open anyway, I don't think it matters if the UI looks awkward too :-)

Implementation note: the main change here is setting
screenOrientation="portrait" in the manifest, but we also need to set
configChanges="keyboardHidden" to prevent the activity from getting
destroyed and recreated when the keyboard slides open or closed.

I'm not getting rid of any landscape-specific code or resources yet; this
is just the AndroidManifest change, so we can do some quick user testing
and gather feedback.

TESTED: On dream, slid keyboard open/closed, verified the UI stayed in
portrait mode, verified (by looking at the logs) that we weren't getting
destroyed or paused.