In-call: don't turn off the screen while DIALING.
This fixes bug 2042741 ("Display blacks out prior to call being
connected").  The problem happens because the CDMA "origination" state is
sometimes very slow, especially with poor coverage, and and it's confusing
for the screen to go black before the call gets connected (or fails.)

The fix: force the screen on the entire time we're in the DIALING state,
just like we already do for the RINGING state.

Note we do *not* do this for the ALERTING state, since on GSM phones
sometimes an outgoing call can stay in the ALERTING state even after the
call gets connected.  (This usually happens with automated voice systems.)

Also note there's a separate but related issue that we shouldn't be using
that 15-second timeout on prox sensor devices in the first place (see bug
2042741.)  I plan to fix that too, but wanted to check this piece in first
since it's the "real" fix for this bug.