Minor UI updates for DISCONNECTING state.
Bug 1409831: display the text "Hanging up" when in the new DISCONNECTING
state.  (Otherwise the DISCONNECTING state is treated just like ACTIVE;
see bugs 1409831 and 1659876 for details.)

Also, in InCallScreen, make sure to null out dialogs at the very end of
onDestroy().  (Not strictly needed, since we already did it in onPause(),
but still good style.)