Add a config resource to control the "BACK key to reject" feature.
Currently, while an incoming call is ringing, BACK behaves just like
ENDCALL: it stops the ringing and rejects the current call.

But this is a bad idea on devices where the BACK key is a capacitive
button, since it's too easy to press the key accidentally as you pull the
phone out of your pocket.

There's no way for an app to tell whether the current device has
capacitive buttons or not, so instead this change adds a new config
resource called "allow_back_key_to_reject_incoming", which is true by
default (to preserve the current behavior), but can be overridden via the
resource overlay for products that don't want this behavior.