Work around "bogus orientation changes" bug with incoming call.
This is a workaround for bug 2081075: Phone not able to receive MT call
when slider is open and display's off.

On the XXXXXXX device, if you receive an incoming call while the sliding
keyboard is open and the screen is off, the InCallScreen gets an endless
stream of config changes and gets destroyed and restarted over and over
till it runs out of memory.

This happens regardless of whether or not the device is in the dock, and
regardless of whether or not the "Switch orientation automatically when
roating phone" setting is on or off.

I don't know why those config changes are happening (and there's probably
a framework bug there somewhere), but for now we can work around the
problem by adding "orientation" to the android:configChanges attribute for
the InCallScreen.

(I'll still leave bug 2081075 open as a framework issue.)