Modified the CALL's extras to provide the application's package name.
The name of the provider is passed using the new GATEWAY_PROVIDER_PACKAGE extra.

Made the gateway number extra a URI and renamed to GATEWAY_PROVIDER_URI to be
more generic and support VOIP.

Renamed copyPhoneProviderExtras to checkAndCopyPhoneProviderExtras
since the copy does not happen if some info is missing.
Added new helper method to detect if an intent has the required
phone provider extras.

Bug: 1984556