Fix issue 2094757: Wired headset connection while in call disables speaker button but leaves speaker on.
The fix consists for one part in never disabling the speaker button: the user has the possibility to enable/disable the speaker even if a headset is connected.
When a headset is connected while the speaker is on, the speaker is automatically turned of but its current state is remembered so that it can be reapplied if the headset is disconnected.