Hide the dialpad "drawer handle" on devices that don't need it.
If this device already has an onscreen "Dialpad" button, don't ever show
the drawer handle (which will be mostly covered up by the buttons anyway.)

This is a temporary cosmetic fix.  The real fix, coming soon, is that we
shouldn't use the SlidingDrawer AT ALL on devices with onscreen buttons.

Also, re-enable verbose logging of the telephony state when entering the
InCallScreen, to track down call state problems on CDMA devices.
Update PhoneUtils.dumpCallState() to show the CdmaPhoneCallState too.