Fix CDMA 3 way calling state.
Again, we get updatePhoneState when the PreciseCall state changes.
This is different from Phone.State which is basically IDLE, OFFHOOK,
RINGING. We check if the previous state in case of GSM/CDMA i.e mPhoneState
and newState in the function updatePhoneState. 3 way calling for CDMA
needs to be handled as a special case, since the network doesn't
report all states. Hence the class CdmaPhoneCallState. Add a similar check
for CDMA3 way calling  to check for previous CdmaPhoneCallState before
sending the CIEV responses. Also, abstract some code to a function.
And change the handler function name so there is no confusion in
the future.