EmergencyDialer: add onscreen Dial button (bug 2093523)
This is basically the same as change 20087, which added a Dial button to
the regular TwelveKeyDialer UI.

(Yes, unfortunately there's a lot of code duplicated between the
TwelveKeyDialer and the EmergencyDialer.  But since one's in apps/Contacts
and the other is in apps/Phone, there's no easy way to share common code.
There's already a TODO comment about this.)