InCallScreen does not wakeup on MT calls
For Private and Unknown Incoming calls, we do not start Contact query, instead
onNewRingingConnection() immediately calls PhoneUtils.showIncomingCallUi()
which causes acquire of FULL Lock. After acquiring FULL lock, end of
onNewRingingConnection() was trying to acquire PARTIAL lock and this
PARTIAL lock acquiring process was releasing the FULL lock, causing
Incoming UI to not show and not wake up the phone. So, to resolve this,
we move acquire of PARTIAL lock to statement before calling StartIncomingCallQuery()
in onNewRingingConnection()