Centralized the processing of the intent's extras in internalResolveIntent.
Previously, the extras were processed in the methods internalResolveIntent and placeCall.
The issue was that if data was missing to build the badge, we would still redirect
the call but don't display the badge.

Now we extract and check all the provider's data as early as possible in internalResolveIntent
and use some fields to store the values collected.

Bug: 1984556