Fix issue 2115199: Speaker key remains on after ending call when wired headset is unplugged.
The problem comes from the fact that when a headset is connected while on call and speaker phone is on, the speaker phone is automatically turned off, but the user requested state is saved by PhoneUtils to restore it when the headset is unplugged.
When the call is terminated, current code in CallNotifier.resetAudioStateAfterDisconnect() will not turn speaker off because the speaker state was already off. When the headset is unplugged later on, the headset disconnect event handler in PhoneApp will restore previous speaker state as saved in PhoneUtils and the speaker will be enabled.

The fix consists in always turning speaker off in CallNotifier.resetAudioStateAfterDisconnect(), regardless of actual speaker state. This makes sure that the speaker state as requested by user is reset by PhoneUtils when a call is terminated.
The headset disconnect event handler in PhoneApp is also modified to ignore headset connection events when the phone is not off hook.