Update patch set 1
Patch Set 1: Fails; I would prefer that you didn't submit this

Unfortunately this change doesn't merge cleanly on the latest master and therefore can't be submitted as is. If it's still relevant, please resolve the conflicts and upload a replacement.

Furthermore, I question the usefulness of such a feature given that it requires to add a preference, and I'd like to hear more in that area.

Finally, a quick glance through the code suggests that there are two independent checkboxes (2G-only and 3G-only) whereas logically those should be mutually exclusive.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=-1