Set Audio Mode to IN_CALL when Phone state changes to OFFHOOK
If Audio Mode is not IN CALL, then set the Audio Mode.  This
changes is needed because for one of the carrier specific test
case, call is originated from the lower layer without
using the UI, and since calling does not go through
DIALING state, it skips the steps of setting the Audio Mode.

Change-Id: I007ffc253611d5675f7102416a42d57f710ff603