Clean up "dialer drawer" usage in the InCallScreen.
Stop having the InCallScreen and OtaUtils directly manipulate the
SlidingDrawer widget that contains the DTMF dialpad.  Instead, have the
DTMFTwelveKeyDialer provide an abstract wrapper around it.  Also, the
DTMFTwelveKeyDialer now finds the SlidingDrawer widget itself, rather than
having it passed in to the constructor.

This way the rest of the phone app doesn't have to know the
SlidingDrawer's resource ID, or care whether or not it even exists.
(This is nice, given that I'm about to remove it on some platforms.)

TESTED:  Ran the full DTMF Dialpad test suite:
  - Type DTMF digits on the onscreen dialpad:
      - tones get sent correctly
      - digits get appended to the digits display
  - Type DTMF digits on the hard keyboard while dialpad is up
      - tones get sent correctly
      - digits get appended to the digits display
  - Type DTMF digits on the hard keyboard while dialpad is hidden
      - tones get sent correctly
      - digits get appended to the digits display correctly even though
        you can't see it.  (They're there if you open the keypad.)
  - Type DTMF digits on the hard keyboard before *ever* bringing up the
    dialpad
      - tones get sent correctly
      - digits get appended to the digits display correctly even though
        you can't see it.  (They're there if you open the keypad.)
  Ran these on:
    - A GSM device
    - A CDMA device
    - A CDMA device during the OTA call (but note there's no
      onscreen "digits" display)