Correct the mute state for new incoming connections
When we receive a new incoming connection for CDMA, we want to make sure
that the audio is automatically unmuted (which is particularly important
when receiving a call-waiting call, but even for normal MT calls). Also
changed scope of an InCallScreen API to allow the onscreen widgets to be
updated when a wired headset is being used to answer/change mute state of
a call.

Bug: 2115195
Change-Id: I2921435e63bd42f285f09028779996f6d8de48de