Set up SCO channel as soon as we ring, instead of when we answer.
SCO setup can take up to 2 seconds with some headsets, so if we setup on answer
the BT headset can miss 2 seconds of audio.

The HFP 1.5 spec allows SCO setup at any time during the in-call process. I am
still somewhat concerned with compatability, so we will have to watch out for
regressions. I tested with the following devices and found no problems:
Sony HBH-IV840
Jabra BT-8010
LG HBS-200
Motorola HS270
Motorola H3

Note this change does not turn on in-band ringing. We set up the SCO channel
when ringing, but do not route audio until the call is answered.

Change-Id: I551ce239393be71d2f6ff45928526d4febff868c