Change the background of the incall dialpad to blue when bluetooth is in use.
InCallScreen: When drawing the in call dial pad, checks if the bluetooth is enabled
and switch the background colors of the keys to blue. Otherwise the keys have
a green background.

DTMFTwelveKeyDialerView: Added a new method to set background of all the keys.

ButtonGridLayout: Added a new method to set all the childrens background (buttons)
to a given drawable.

New assets for the blue background for the keys and associated selector.

Bug: 2136523