wired Head Set do not mute emg call.
The issue this patch fixes is:
On an emergency call with a wired headset attached, the in-call screen
mute key is disabled but pressing send/end key on the wired head set
mutes the call, but UI does not reflect the mute status.

The fix is to check for emergency call before changing the mute status.

Change-Id: I64c36fe826bd63632dd9d7f33106547bcf755193