Pass the background for the dialer's button using resource id.
Previously, we passed it as a drawable and this produced some
glitches in the way the buttons were rendered.
I thought using a drawable was going to be more efficient.
Will log a but with the UI to get an explaination why
I had to make that change.

This time I actually punched numbers instead of just testing the
background change based on the BT connectivity.