Refactor Cdma OTASP InCallScreenState to be saved in PhoneApp via OtaUtils.
State that needs to outlive the life of an activity must reside
in an application. This fixes a bug which the caused the OTASP
to apparently not complete because if the Activity was recreated
then last OTASP state would be lost.

One way to see this occur is when in OTASP mode press the power
key to turn off the display which makes the InCallScreen Activity
a background task. When InCallScreen moves to the foreground it
is recreated and we lose state and there is no Next button displayed.

Bug: 2133182
Change-Id: Ife9753bf7689b48be4c5e78f95be95250d0a55b5