Align the buttons in the dialer at the bottom of their view.
Previously we could not use layout_weight to distribute the free space
between the dialpad and the button row in the emergency dialer.
The reason was that the dialpad's buttons were layed out starting at the top left corner.

Now ButtonGridLayout position the buttons so that the last row is at
the bottom of the view (+padding).
Similarly the row of vm/call/bspace buttons is now top aligned in its view so we
can control the space between the last row of button in the dialer
and the vm/call/backspace row of buttons accurately and let the layout_weight
attributes distribute the remaining space between these 2.
As are result the hardcoded margin are gone so this layout should work on
WVGA 854 and WVGA 800.

Tested on WVGA 854 and HVGA.