Fixed the cut buttons in the incall dialpad.
There was mistake in the calculation of the first row positioning.
I forgot to add the vertical padding hence the buttons were positioned
too high in the view.
As a result the paddings of the 'torpedo' had to be adjusted too.

Fixed the button size in the hvgs non drawer dialpad.
Created a new long version of the non drawer dialpad

In the drawer version of the dialpad, added a dummy element that
consumes some of the empty space below the dialpad, in effect
pushing it up (it was stucked at the bottom of the drawer).

Bug:2160234