Removed the explicit image view for the provider's icon.
Instead of using an explicit element to position the image,
TextView has the capability of positioning Drawable left of
text being displayed.
Also lifted the restriction on the number of lines in case
an app with a long label (or the German version) needs
more space to display the message.

Bug: 2170873