Fix audio hiss on call end scenarios
When ending a call, we hear the call tone, but there is also a hiss
that we hear until the ToneGenerator is finally released and
destroyed. Therefore, change the timeout value for these various
call end tones to be the same as the actual duration of the tone
we are playing.

Change-Id: I954352aaa2ec3a7e1404526e0149566bcf6fd995
Bug-Id: 2173790