Bluetooth A2DP suspend-resume improvements.
This change will reduce the occurence rate of A2DP sink suspend resume failures observed in issues 2184627, 2181005 and possibly 2189628.

Make sure the A2DP suspend flag is restored when bluetooth is disabled or BT device is disconnect.
Do not attempt to resume the A2DP sink if the BT device has been disconnected.