Allow the photo update when the call card is in DIALING state.
Currently the call card in DIALING state displays the generic
dialing photo. In updatePhotoForCallState we exclude that state
from the update. I just reenabled it by moving it before the
default case with does the work.
Instead of removing it I left the case statement to make explicit
that in Dialing state we are hitting the code in default.

Tested in normal and 3 way scenarios.

Bug: 131379