3-Way calling dialing UI bug fix for CDMA
Issue: For CDMA, on making a 3way call, sometimes the 2nd called party number
update is incorrect (when the called number is *not* present in Contacts).

Fix: For CDMA, we need to be using call.getLatestConnection() for getting the
latest connection info in the forground call(as there is no concept of background
call in CDMA).  In CallCard.onQueryComplete() we now use call.getLatestConnection()
instead of the usual call.getEarliestConnection().

Note: For GSM, call.getEarliestConnection() is still valid.

Change-Id: I3e89069cbd77b629f26350b96e3a2ec514846f19