Fix BT PTS certification case for GSM devices.
When a call is active, and another call is made from the headset,
we don't respond with OK to command when the call is being dialed.
The PhoneState hasn't changed but the foreground call state has
changed from Active to Dialing.
Fix in discussion with: Fan Jackson who reported it