FIX 2139254: Get rid of "Manage Conference" button, use the photo instead
If the foreground call is a GSM conference call, replace the usual contact
photo with a big clickable button (with basically the same icon as before)
that brings up the Manage Conference UI.

This allows us to get rid of the original "Manage Conference" button at
the upper-right of the screen, which could collide with the "on hold"
photo.

TESTED: GSM conference calls, hvga and wvga.  Verified this change has no
effect on CDMA devices *or* devices like Dream/Sapphire with no onscreen
touch UI in the phone app.