Removed the code that tries to approximate the user touch event.
Matching the user touch event with a button should be done by the
framework, we should not have to write custom code to do so.
Also that code is broken because it tries to approximate any touch
event to a key which is wrong. We could fix it by limiting the matching zone
but even then it is not clear it would work. For instance if a layout make
the buttons touch, the matching zone could overlap which would not work
properly.

I am removing that code (dispatchTouchEvent) and will ask QA to test it
thoroughly to check if the in call dialpad lost some accuracy.

Bug:2113207