Update patch set 1
Patch Set 1: No score; Do not submit

Sorry, I realize it's bad that we don't (yet) provide a way for 3rd party
apps to do this kind of stuff, but it doesn't make sense to add API to the
InCallScreen  for this.

The InCallScreen is purely the *UI* for the in-call experience, but
functions like "answer" / "conference" / "hold" / etc. really should go
directly to the telephony layer, which can handle them without involving
the Activity system at all, and which also can protect them with
appropriate persmissions.

(FWIW, you *can* use the current ITelephony interface for a couple of
these right now, at least "answer" and "hang up".  but that's not a public
API yet so there's no guarantee it'll still be there in upcoming
releases.)

We do plan to update the telephony APIs ourselves (probably via
android.telephony.TelephonyManager) and (re)expose them to SDK developers,
but I just can't say for sure what the schedule is for that.

Patch-set: 1
Reviewer: David Brown <1000358@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0