In getCompactNameFromCallerInfo, don't call modifyCnapCases on a name.
When the async query to lookup a CallerInfo completes, we update the logs
with a compact name which is the name of the contact associated with
the number or the number itself.
If we have a name, we should not call modifyForSpecialCnapCases because
it takes a number a argument. Special number strings are used to
mark a call as private and it clashes with some contact name entries.

For instance, a contact with name 'P' will always be marked as private
number even thought it is not.

Bug:2160795