Display the unknown photo icon while we hit the contact DB.
The method that makes async calls to the contact DB to retrieve a picture,
takes a resource id which is the image to be shown while the call runs.

Currently we use the value -1 which means that the image element will be non
visible. As a result the picture is hidden. When the DB is non responsive, the query
takes a long time and the user sees a call card with no picture.

To work around that, we now pass the resource id of the unknown contact picture,
which gets displayed.

Bug:2239852