Fix an obscure scenario where a single CALL request would cause two outgoing call attempts.
Background: OutgoingCallBroadcaster is an invisible activity which receives
the CALL intent (from the dialer, contacts, etc.) and translates it into a
special broadcast called NEW_OUTGOING_CALL, which ultimately launches the
in-call UI.

Unfortunately if an orientation change happened at the exact right moment, the
ActivityManager would destroy and recreate the OutgoingCallBroadcaster, which
resulted in a *second* separate NEW_OUTGOING_CALL broadcast, which resulted in
two separate CALL intents coming in to the InCallScreen.  In some phone
states, that 2nd call would fail, causing an error dialog to come up saying
"Call not sent".  See bug 2202413 for the full analysis.

There are two parts of the fix:

(1) Fix OutgoingCallBroadcaster to gracefully handle the situation where it
    gets killed and re-created, i.e. when icicle != null in onCreate().
    In that case we just do nothing, since the NEW_OUTGOING_CALL broadcast was
    already sent the first time that onCreate() ran.

(2) Set android:configChanges="orientation|keyboardHidden" in the manifest for
    the OutgoingCallBroadcaster activity.  This prevents the ActivityManager
    from killing and restarting us in the first place.

Note that either part by itself actually fixes this bug (and I did test them
separately), but they're both technically correct, so this change contains
both.

BUG=2202413 ('call not sent' prompt displayed when we add call in landscape mode)

TESTED:
  Both GSM and CDMA:
    - Making an outgoing call from contacts
    - Making an outgoing call from the dialer
    - Making an outgoing call from the call log
    - Making an outgoing call from the emergency dialer
    - Making a 2nd outgoing call while on a call
  CDMA only:
    - Sending an empty flash (pressing DIAL in the dialer without any
      digits, while a call is already in progress)