Fix 2209086: Make animation/decline transition much smoother in PhoneApp.
This fix makes hideIncomingCallWidget() re-entrant such that repeatedly calling it while
the incoming call widget is being hidden is OK. This fixes a problem where we might try to
hide the incoming call widget a second time in the case where updateState() is called again
based on some async event.

Tested: Answer/Decline call on CDMA and GSM device.