Remove use of the FLAG_DISMISS_KEYGUARD window manager flag
This flag was added to allow the user to go directly to the home screen
when pressing the Home button while in-call and the keyguard does not have a lock pattern set.
But unfortunately this introduced a bug when an incoming phone call is declined.
In that case, the insecure keyguard would be automatically dismissed, which could result in
the phone staying on while in your pocket.

Change-Id: Ie7d3423ca50d093c6f4fd4049d9965eeaec2a235
Signed-off-by: Mike Lockwood <lockwood@android.com>
Bug: 2330359