Use the FLAG_DISMISS_KEYGUARD while we are in-call.
This is a better fix for bug b/2330359.
When we are actively in-call, we now set the FLAG_DISMISS_KEYGUARD flag
so the user can press Home to exit directly to the home screen.
But in the case of a declined or unanswered phone call we will drop into
the keyguard after the call terminates.

Change-Id: I6ee0759dbedeaf05e940558ed27f09f58c6582fa
Signed-off-by: Mike Lockwood <lockwood@android.com>
Bug: 2330359