Fully populate the AccessibilityEvent for the in-call UI.
This is a patch from Svetoslav Ganov <svetoslavganov@google.com> from the
Accessibility team.

Since the "call card" component of the in-call UI has no focusable
elements, that causes the "TalkBack" accessibility feature to ignore some
important info like the name and number of who's calling you :-(

The fix is for the CallCard to manually fill in the AccessibilityEvent
based on the contents of the various onscreen UI elements.

BUG=2375776
TESTED=verified in-call UI is read correctly when TalkBack is enabled.
DRNO=timsullivan