Hangup active call not ringing call on AT+CHUP.
Plantronics @ Unplugfest recommended to hangup active call in the active/incoming case. The HFP spec does not clearly say what to do here, it just says that CHUP can hangup incoming call or active call. The 3GP spec does say to hangup the current call. So it seems like Plantronics are correct to hangup the active call in the active/ringing case.

Change-Id: I973eb607b46cd83351d0ea1513b699bc032e30b1
Signed-off-by: Nick Pelly <npelly@google.com>