Use mNumber to decide whether to add/edit an FDN entry.
Look at mNumber instead of mName when deciding whether to add
or edit a fixed dialling list entry

3GPP TS 31.102 (USIM application) v9.1.0 states in chapter
4.2.24 EFFDN (Fixed Dialling Number) that the Dialling Number
field is mandatory whereas the Alpha Identifier is optional.
The current implementation of the user interface for editing
the fixed dialling list blocks creation and deletion list
entries without a name (corresponding to the Alpha Identifier)
but instead permits entries without a number. This is incompatible
with the USIM specification and implementation. This contribution
corrects the implementation in line with the USIM specification.

Change-Id: Ibf364207b89e40e524e18d5056cc2aa1abfabe0e