Fix issue 2413494: Add TTY support.
The TTY implementation is modified in order to operate on several platforms:
- The actual TTY mode configuration is moved from CallFeaturesSettings to PhoneApp.
CallFeaturesSettings only updates the user selected preferred TTY operating mode
in the Settings data base and broadcasts an intent indicating the preferred mode
change.
- Phone app processes this intent and only applies the user requested mode if a headset
is connected: The new TTY mode is indicated to both RIL and audio hardware.
- A consequence is that the TTY indication in the status bar is only shown if TTY mode is
not OFF AND the RIL has accepted the TTY configuration AND a device is connected.