Don't block in the main thread for an outgoing connection to cancel.
We cancel the outgoing connection when we receive an incoming connection at the
same time. I added the code to wait for the outgoing thread to complete out of
paranoia. Looking at the outgoing thread, I don't think there is any harm if we
let it complete asynchronously. The interrupted code paths do not change any
global state.

Bug: 2315488
Bug: 2203159
Change-Id: Ie280aa93bf883e357950d13e79ef35842da7c739