Fix 2432289: passion#c: no DTMF feedback during activation
Local playback of DTMF tones would work fine during a normal call, but not
during the OTA call.  This was because the OtaUtils class was creating its
own "DTMFTwelveKeyDialer" instance, but never properly initializing it.

This bug has always been there, but we never noticed this on Sholes,
because local DTMF tones are played automatically by the Sholes radio.
(So on Sholes that DTMFTwelveKeyDialer instance didn't actually *need* to
make any sound.)

The fix is to correctly initialize the DTMFTwelveKeyDialer instance that
we create: we call startDialerSession() on it from initOtaInCallScreen(),
and call stopDialerSession() from cleanOtaScreen().

It's still pretty confusing how we have two separate DTMFTwelveKeyDialer
instances around, but we're very close to eclair code freeze so I'm
keeping this change as minimally invasive as possibile.  (I added a TODO
comment about eventually cleaning this up, though.)

Tested: On PassionC, local tones are now audible during the OTA call.
  (They're pretty faint unless you turn the speaker on, but I think that's
  normal, or possibly due to bug 2447744.)

Bug: 2432289
DrNo: hiroshi