Fix crash in TimeConsumingPreferenceActivity on back key press
Fix issue 2483803:  Crash on back key press during "Reading call settings"
message dialog.

Add state to track foreground status of activity and only dismiss dialog
if activity is foreground.  Also cover a couple of other cases in this
class where it could attempt to show/dismiss a dialog when not in the
foreground beyond the scope of this bug.