In the ALERTING state, don't replace the image with the dialing one.
When making calls using a GSM phone, there is an ALERTING state,
after the DIALING one. Currently in that state we replace the picture
image with the generic green dialing icon.

This change removes that and the picture is left alone. Note that
the picture could be a custom picture or the generic Android face
if the contact does not have one.

A drawback with this change is that we lost the ALERTING indication.
We could replace the string 'Dialing' with 'Ringing' if we wanted
to inform the user of the call's progress.

Bug:2479193