Fix phone crash in BT MT call disconnect, enable mute/hold in GSM.
When phone is in global mode and acquires CDMA signal first, it will start an
OTA call. OTA variables get initiazed which are only valid in CDMA. On switching
to GSM/UMTS these variables are not cleared. MUTE/HOLD options in InCallMenu are
enabled/disabled based on these variables and hence MUTE/HOLD options are
disabled for normal GSM calls. Similarly while disconnecting the MT call in BT,
the OTA vars are checked and some processing is done which leads to phone crash.

To fix these issues, phone type is checked for CDMA before using OTA data. OTA
variables are cleaned up and call screen is cleaned up when there is a
radio technology change.