Allow PIN/PUK related "special sequences" in the Emergency Dialer.
Way back in change https://android-git.corp.google.com/g/30821 we disabled
"special sequences" in the Emergency Dialer (see bug 2168271).

But it turns out you *do* actually need to use some sequences in the
Emergency Dialer, like the PIN/PUK related "**04" or "**05" codes.
(These allow you to reset your PIN if you're locked out of your phone, and
the only way to do that is via the Emergency Dialer.)

So the Emergency Dialer now allows *only* the subset of special sequences
that need to be available even if the device is locked.  (Those sequences
are handled by the new SpecialCharSequenceMgr handleCharsForLockedDevice()
method.)

Bug: 2368835

Tested (on Passion):
  - From emergency dialer (with device locked):
    - "*#06#" does NOT display the IMEI
    - "**05<blah>#" codes *are* allowed

  - From regular dialer (with device unlocked):
    - "*#06#" displays the IMEI
    - "**05<blah>#" codes are still allowed

Change-Id: Ie4fdd35ac9142ad33a7339080c9e668b0f584974