Fix 2515687: Cannot dial emergency number while in no service
Fix logic in InCallScreen.placeCall() to make sure we allow emergency
numbers in both EMERGENCY_ONLY mode *and* also the OUT_OF_SERVICE state.

(Even if we're OUT_OF_SERVICE, we still attempt to make a call, since the
radio will register to any available network.)

I had accidentally broken the OUT_OF_SERVICE case with change
https://android-git.corp.google.com/g/42697.  That change was in master
(on Mar 2, 2010) so this fix is only necessary in froyo.

Bug: 2515687

Change-Id: I3d8f612a87d5986084d07dc94551459b7d1147f8
Tested: on Sholes (by using the Extended NAM menu to add a fake
    Emergency number, and manually hacking the result of
    mPhone.getServiceState() to be OUT_OF_SERVICE.)