Minor cleanup of checkIsOtaCall() method
Update InCallScreen.checkIsOtaCall() to bail out instantly if this isn't a
CDMA phone, to make it perfectly clear that the following code is all
CDMA-specific.  Clarify this in the javadoc too.

(There's no functional change here, since on GSM phones we'll bail out
anyway a few lines later when we notice app.cdmaOtaScreenState is null.
This extra check just makes the code more clear.)

Change-Id: I28b2bcb399edf3268f4212ed9aeda91e1421df22